fix
long and int build problem

+++ src/main/java/com/gni/banking/Service/UserService.java
@@ -125,7 +125,7 @@ public class UserService {
     }
 
 
-    public double getDayLimitById(int userId){
+    public double getDayLimitById(long userId){
         return userRepository.getDayLimitById(userId);
     }
